Adding IExporter and IExporterFactory interfaces; refactoring CSV export to dedicated CsvExporter class
