+++ src/Endatix.Infrastructure/Builders/InfrastructureDataBuilder.cs
@@ -2,10 +2,12 @@ using Endatix.Core.Abstractions;
 using Endatix.Core.Abstractions.Data;
 using Endatix.Core.Abstractions.Exporting;
 using Endatix.Core.Abstractions.Repositories;
+using Endatix.Core.Entities;
 using Endatix.Core.Infrastructure.Domain;
 using Endatix.Infrastructure.Data;
 using Endatix.Infrastructure.Data.Repositories;
 using Endatix.Infrastructure.Exporting;
+using Endatix.Infrastructure.Features.Submissions;
 using Endatix.Infrastructure.Repositories;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -46,7 +48,8 @@ public class InfrastructureDataBuilder
         Services.AddScoped(typeof(IRepository<>), typeof(EfRepository<>));
         Services.AddScoped<IFormsRepository, FormsRepository>();
         Services.AddScoped<ISubmissionExportRepository, SubmissionExportRepository>();
-        Services.AddScoped<IExporter, CsvExporter>();
+        Services.AddScoped<IExporterFactory, ExporterFactory>();
+        Services.AddScoped<IExporter<SubmissionExportRow>, SubmissionCsvExporter>();        
         Services.AddSingleton<DataSeeder>();
 
         LogSetupInfo("Data infrastructure configured successfully");
